requirements.txt
streamlit-summarizer-app/streamlit-summarizer-app/requirements.txt

streamlit
transformers
torch
PyMuPDF
Pillow
pytesseract
streamlit-ace